Image caching
If an image has already been generated before then it is reused instead of being generated again, saving on generation time. For instance, a video with 2 seeds that loops back to the start will finish in half the time, and using compare_paths won't need to re-generate the first image every time. The total number of images generated is now only the number of unique images.